refactor: decompose 3 god files below 700 loc
- errorHandling.ts: 828 to 413 (extract artistNameFilter.ts)
- IntegratedStemTracks.tsx: 837 to 694 (extract stemTrackConfig.ts)
- LyricsParser.ts: 742 to 558 (extract lyricsAnalysisSuite.ts)
- 904 tests passing, backward-compatible exports
